requirements.txt
torchio>=0.17.46
matplotlib>=3.1.1
nibabel>=2.5.1
numpy>=1.19.0
scikit-learn>=0.23.1
SimpleITK==1.2.4
gdown>=3.11.1
